requirements.txt
PySimpleGUI==4.60.4
requests==2.28.2
GitPython==3.1.31
https://github.com/pyinstaller/pyinstaller/archive/develop.zip